# backend/src/routes/prompts.js
/**
 * Prompts Routes
 * Handles CRUD operations for user prompts
 */

import express from 'express'

const router = express.Router()

/**
 * @route GET /
 * @desc Get all prompts for the authenticated user
 * @access Private
 */
router.get('/', (req, res) => {
  res.status(200).json({ message: 'Get all prompts endpoint - Not Implemented' })
})

/**
 * @route GET /:id
 * @desc Get a specific prompt by ID
 * @access Private
 */
router.get('/:id', (req, res) => {
  res.status(200).json({ message: `Get prompt ${req.params.id} endpoint - Not Implemented` })
})

/**
 * @route POST /
 * @desc Create a new prompt
 * @access Private
 */
router.post('/', (req, res) => {
  res.status(201).json({ message: 'Create prompt endpoint - Not Implemented' })
})

/**
 * @route PUT /:id
 * @desc Update a prompt by ID
 * @access Private
 */
router.put('/:id', (req, res) => {
  res.status(200).json({ message: `Update prompt ${req.params.id} endpoint - Not Implemented` })
})

/**
 * @route DELETE /:id
 * @desc Delete a prompt by ID
 * @access Private
 */
router.delete('/:id', (req, res) => {
  res.status(204).json({ message: `Delete prompt ${req.params.id} endpoint - Not Implemented` })
})

export default router 